added

+++ lib/models/item_saved_model.dart
@@ -6,7 +6,7 @@ class ItemSavedModel {
   late List<String> dataMap;
   late double qouted;
   late double addenum;
-  late int count;
+  late double count;
   late int baseId;
   late double cost;
   ItemSavedModel();
